Skip checking kernel vintf requirements
Inorder for kernel vintf requirements to succeed, target level needs to
be set to 8 and kernel requirements needs to be met.

As the HAL requirements are not set, target level can't be set to 8.
If PRODUCT_OTA_ENFORCE_VINTF_KERNEL_REQUIREMENTS is set to true, build
will fail due to requirements not met.

Fix the issue by setting PRODUCT_OTA_ENFORCE_VINTF_KERNEL_REQUIREMENTS
to false.

Tracked-On: OAM-111022
Signed-off-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>